packages/unxt-api/pyproject.toml
[project]
authors = [
  { name = "GalacticDynamics", email = "nstarman@users.noreply.github.com" },
  { name = "Nathaniel Starkman", email = "nstarman@users.noreply.github.com" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python",
  "Topic :: Scientific/Engineering",
  "Typing :: Typed",
]
dependencies = ["astropy>=7.0.0", "plum-dispatch>=2.5.7"]
description = "Abstract dispatch API definitions for unxt"
dynamic = ["version"]
license.text = "BSD-3-Clause"
name = "unxt-api"
readme = "README.md"
requires-python = ">=3.11"

[project.urls]
"Bug Tracker" = "https://github.com/GalacticDynamics/unxt/issues"
Homepage      = "https://github.com/GalacticDynamics/unxt"

[build-system]
build-backend = "hatchling.build"
requires      = ["hatch-vcs", "hatchling"]

[tool.hatch.version]
raw-options = { root = "../..", search_parent_directories = true, git_describe_command = "git describe --dirty --tags --long --match 'unxt-api-v*'" }
source      = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/unxt_api/_version.py"
